xlate: Clear conntrack fields when traversing peers.
When traversing from one bridge to another via a patch port, we should
drop conntrack metadata from the OpenFlow pipeline. This should only
affect the subsequent processing on the destination bridge (and not
continued processing, resubmits, etc. on the original bridge).

If conntrack is executed for the current bridge, then recirculation
should make the conntrack fields available for matching. In all other
cases, the conntrack fields should be zeroed during upcall processing.

CC: Jarno Rajahalme <jrajahalme@nicira.com>
Signed-off-by: Joe Stringer <joestringer@nicira.com>
---
This should be folded into earlier conntrack patches, but I have left it
as an independent patch for now to aid review.