Update Mark/label generation for newer kernel API.
XXX: Merge this into earlier patches.

Recent feedback for the kernel datapath has requested that the mark and
label be set as part of the conntrack action. To retain the current
OpenFlow interface, this patch allows the commit of conntrack actions to
be deferred as late as possible, so that OpenFlow's
"ct(commit),set_field:1->ct_mark" can be turned into datapath
"ct(commit,mark=1)".