XXX packets: Do not zero all ct_* metadata.
introducing new members in pkt_metadata can have a negative impact on
the throughput of DPDK datapath, because the structure needs to be
initialized for each packet.

By making the assumption that ct_* member are only used if ct_state is
non zero, we can avoid zeroing out some bytes for each packet.

The same logic has been used for the tunnel metadata.